+++ crates/koto/src/lib.rs
@@ -34,6 +34,6 @@ pub mod prelude;
 pub use koto_bytecode as bytecode;
 pub use koto_parser as parser;
 pub use koto_runtime as runtime;
-pub use koto_runtime::{derive, Borrow, BorrowMut, Error, Ptr, PtrMut, Result};
+pub use koto_runtime::{derive, Borrow, BorrowMut, Error, ErrorKind, Ptr, PtrMut, Result};
 
 pub use crate::koto::{Koto, KotoSettings};